+++ lib/Query/QueryProvider.php
@@ -2,7 +2,7 @@
 /**
  * Nextcloud - user_sql
  *
- * @copyright 2018 Marcin Łojewski <dev@mlojewski.me>
+ * @copyright 2020 Marcin Łojewski <dev@mlojewski.me>
  * @author    Marcin Łojewski <dev@mlojewski.me>
  *
  * This program is free software: you can redistribute it and/or modify
@@ -54,6 +54,25 @@ class QueryProvider implements \ArrayAccess
     }
 
     /**
+     * Get first nonempty property value.
+     *
+     * @param mixed ...$keys Property keys.
+     *
+     * @return mixed
+     */
+    private function firstNonemptyProperty(...$keys)
+    {
+        foreach ($keys as $key) {
+            $value = $this->properties[$key];
+            if (!empty($value)) {
+                return $value;
+            }
+        }
+
+        return null;
+    }
+
+    /**
      * Load queries to the array.
      */
     private function loadQueries()
@@ -64,18 +83,19 @@ class QueryProvider implements \ArrayAccess
 
         $gAdmin = $this->properties[DB::GROUP_ADMIN_COLUMN];
         $gGID = $this->properties[DB::GROUP_GID_COLUMN];
-        $gName = $this->properties[DB::GROUP_NAME_COLUMN];
+        $gName = $this->firstNonemptyProperty(DB::GROUP_NAME_COLUMN, DB::GROUP_GID_COLUMN);
 
         $uActive = $this->properties[DB::USER_ACTIVE_COLUMN];
         $uAvatar = $this->properties[DB::USER_AVATAR_COLUMN];
         $uDisabled = $this->properties[DB::USER_DISABLED_COLUMN];
         $uEmail = $this->properties[DB::USER_EMAIL_COLUMN];
         $uHome = $this->properties[DB::USER_HOME_COLUMN];
-        $uName = $this->properties[DB::USER_NAME_COLUMN];
+        $uName = $this->firstNonemptyProperty(DB::USER_NAME_COLUMN, DB::USER_USERNAME_COLUMN, DB::USER_UID_COLUMN);
         $uPassword = $this->properties[DB::USER_PASSWORD_COLUMN];
         $uQuota = $this->properties[DB::USER_QUOTA_COLUMN];
         $uSalt = $this->properties[DB::USER_SALT_COLUMN];
         $uUID = $this->properties[DB::USER_UID_COLUMN];
+        $uUsername = $this->firstNonemptyProperty(DB::USER_USERNAME_COLUMN, DB::USER_UID_COLUMN);
 
         $ugGID = $this->properties[DB::USER_GROUP_GID_COLUMN];
         $ugUID = $this->properties[DB::USER_GROUP_UID_COLUMN];
@@ -87,16 +107,18 @@ class QueryProvider implements \ArrayAccess
         $quotaParam = Query::QUOTA_PARAM;
         $searchParam = Query::SEARCH_PARAM;
         $uidParam = Query::UID_PARAM;
+        $usernameParam = Query::USERNAME_PARAM;
 
         $reverseActiveOpt = $this->properties[Opt::REVERSE_ACTIVE];
 
         $groupColumns
             = "g.$gGID AS gid, " .
-            (empty($gName) ? "g." . $gGID : "g." . $gName) . " AS name, " .
+            "g.$gName AS name, " .
             (empty($gAdmin) ? "false" : "g." . $gAdmin) . " AS admin";
         $userColumns
             = "u.$uUID AS uid, " .
-            (empty($uName) ? "u." . $uUID : "u." . $uName) . " AS name, " .
+            "u.$uUsername AS username, " .
+            "u.$uName AS name, " .
             (empty($uEmail) ? "null" : "u." . $uEmail) . " AS email, " .
             (empty($uQuota) ? "null" : "u." . $uQuota) . " AS quota, " .
             (empty($uHome) ? "null" : "u." . $uHome) . " AS home, " .
@@ -115,9 +137,8 @@ class QueryProvider implements \ArrayAccess
             Query::COUNT_GROUPS =>
                 "SELECT COUNT(ug.$ugGID) " .
                 "FROM $userGroup ug " .
-                "WHERE ug.$ugGID = :$gidParam " .
-                "AND ug.$ugUID " .
-                "LIKE :$searchParam",
+                "WHERE ug.$ugGID LIKE :$gidParam " .
+                "AND ug.$ugUID LIKE :$searchParam",
 
             Query::COUNT_USERS =>
                 "SELECT COUNT(u.$uUID) AS count " .
@@ -133,9 +154,8 @@ class QueryProvider implements \ArrayAccess
             Query::FIND_GROUP_USERS =>
                 "SELECT ug.$ugUID AS uid " .
                 "FROM $userGroup ug " .
-                "WHERE ug.$ugGID = :$gidParam " .
-                "AND ug.$ugUID " .
-                "LIKE :$searchParam " .
+                "WHERE ug.$ugGID LIKE :$gidParam " .
+                "AND ug.$ugUID LIKE :$searchParam " .
                 "ORDER BY ug.$ugUID",
 
             Query::FIND_GROUPS =>
@@ -145,16 +165,34 @@ class QueryProvider implements \ArrayAccess
                 (empty($gName) ? "" : "OR g.$gName LIKE :$searchParam ") .
                 "ORDER BY g.$gGID",
 
-            Query::FIND_USER =>
-                "SELECT $userColumns, u.$uPassword AS password " .
+            Query::FIND_USER_BY_UID =>
+                "SELECT $userColumns " .
                 "FROM $user u " .
                 "WHERE u.$uUID = :$uidParam " .
                 (empty($uDisabled) ? "" : "AND NOT u.$uDisabled"),
 
-            Query::FIND_USER_CASE_INSENSITIVE =>
+            Query::FIND_USER_BY_USERNAME =>
                 "SELECT $userColumns, u.$uPassword AS password " .
                 "FROM $user u " .
-                "WHERE lower(u.$uUID) = lower(:$uidParam) " .
+                "WHERE u.$uUsername = :$usernameParam " .
+                (empty($uDisabled) ? "" : "AND NOT u.$uDisabled"),
+
+            Query::FIND_USER_BY_USERNAME_CASE_INSENSITIVE =>
+                "SELECT $userColumns, u.$uPassword AS password " .
+                "FROM $user u " .
+                "WHERE lower(u.$uUsername) = lower(:$usernameParam) " .
+                (empty($uDisabled) ? "" : "AND NOT u.$uDisabled"),
+
+            Query::FIND_USER_BY_USERNAME_OR_EMAIL =>
+                "SELECT $userColumns, u.$uPassword AS password " .
+                "FROM $user u " .
+                "WHERE u.$uUsername = :$usernameParam OR u.$uEmail = :$emailParam " .
+                (empty($uDisabled) ? "" : "AND NOT u.$uDisabled"),
+
+            Query::FIND_USER_BY_USERNAME_OR_EMAIL_CASE_INSENSITIVE =>
+                "SELECT $userColumns, u.$uPassword AS password " .
+                "FROM $user u " .
+                "WHERE lower(u.$uUsername) = lower(:$usernameParam) OR lower(u.$uEmail) = lower(:$emailParam) " .
                 (empty($uDisabled) ? "" : "AND NOT u.$uDisabled"),
 
             Query::FIND_USER_GROUPS =>