+++ lib/Repository/UserRepository.php
@@ -2,7 +2,7 @@
 /**
  * Nextcloud - user_sql
  *
- * @copyright 2018 Marcin Łojewski <dev@mlojewski.me>
+ * @copyright 2020 Marcin Łojewski <dev@mlojewski.me>
  * @author    Marcin Łojewski <dev@mlojewski.me>
  *
  * This program is free software: you can redistribute it and/or modify
@@ -55,22 +55,62 @@ class UserRepository
     /**
      * Get an user entity object.
      *
-     * @param string $uid           The user ID.
+     * @param mixed $uid The user ID.
+     *
+     * @return User The user entity, NULL if it does not exists or
+     *              FALSE on failure.
+     */
+    public function findByUid($uid)
+    {
+        return $this->dataQuery->queryEntity(
+            Query::FIND_USER_BY_UID, User::class, [Query::UID_PARAM => $uid]
+        );
+    }
+
+    /**
+     * Get an user entity object.
+     *
+     * @param string $username      The username.
      * @param bool   $caseSensitive TRUE for case sensitive search,
      *                              FALSE for case insensitive search.
      *
      * @return User The user entity, NULL if it does not exists or
      *              FALSE on failure.
      */
-    public function findByUid($uid, $caseSensitive = true)
+    public function findByUsername($username, $caseSensitive = true)
     {
         if ($caseSensitive) {
             return $this->dataQuery->queryEntity(
-                Query::FIND_USER, User::class, [Query::UID_PARAM => $uid]
+                Query::FIND_USER_BY_USERNAME, User::class, [Query::USERNAME_PARAM => $username]
             );
         } else {
             return $this->dataQuery->queryEntity(
-                Query::FIND_USER_CASE_INSENSITIVE, User::class, [Query::UID_PARAM => $uid]
+                Query::FIND_USER_BY_USERNAME_CASE_INSENSITIVE, User::class, [Query::USERNAME_PARAM => $username]
+            );
+        }
+    }
+
+    /**
+     * Get an user entity object.
+     *
+     * @param string $query         The username or email address.
+     * @param bool   $caseSensitive TRUE for case sensitive search,
+     *                              FALSE for case insensitive search.
+     *
+     * @return User The user entity, NULL if it does not exists or
+     *              FALSE on failure.
+     */
+    public function findByUsernameOrEmail($query, $caseSensitive = true)
+    {
+        if ($caseSensitive) {
+            return $this->dataQuery->queryEntity(
+                Query::FIND_USER_BY_USERNAME_OR_EMAIL, User::class,
+                [Query::USERNAME_PARAM => $query, Query::EMAIL_PARAM => $query]
+            );
+        } else {
+            return $this->dataQuery->queryEntity(
+                Query::FIND_USER_BY_USERNAME_OR_EMAIL_CASE_INSENSITIVE, User::class,
+                [Query::USERNAME_PARAM => $query, Query::EMAIL_PARAM => $query]
             );
         }
     }