+++ CHANGELOG.md
@@ -4,6 +4,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.7.1]
+- Support for Nextcloud 22, 23
+
+## [4.7.0]
+### Changed
+- Support for Doctrine 3
+- Support for Nextcloud 21 only
+
+## [4.6.0] - 2021-01-16
+### Fixed
+- [issue#123](https://github.com/nextcloud/user_sql/issues/123) - sync exception - Duplicate entry
+### Added
+- Support for Nextcloud 20
+
+## [4.5.0] - 2020-04-13
+### Added
+- Support for Nextcloud 19
+- Argon2id support
+- System wide values option
+- Allow email login option
+- UID user table column
+- GID user table column
+- HMAC hash implementation
+- Default group option
+
 ## [4.4.1] - 2020-02-02
 ### Fixed
 - Issue generating 'Object of class Closure could not be converted to string' log message
@@ -138,6 +163,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Supported version of ownCloud, Nextcloud: ownCloud 10, Nextcloud 12
 
+[4.7.1]: https://github.com/nextcloud/user_sql/compare/v4.7.0...v4.7.1
+[4.7.0]: https://github.com/nextcloud/user_sql/compare/v4.6.0...v4.7.0
+[4.6.0]: https://github.com/nextcloud/user_sql/compare/v4.5.0...v4.6.0
+[4.5.0]: https://github.com/nextcloud/user_sql/compare/v4.4.1...v4.5.0
 [4.4.1]: https://github.com/nextcloud/user_sql/compare/v4.4.0...v4.4.1
 [4.4.0]: https://github.com/nextcloud/user_sql/compare/v4.3.0...v4.4.0
 [4.3.0]: https://github.com/nextcloud/user_sql/compare/v4.2.1...v4.3.0